win_backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==2.0.0
colorama==0.4.6
Django==4.1.3
djangorestframework==3.14.0
exceptiongroup==1.0.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
pycodestyle==2.9.1
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
